+++ cx-portal/src/features/store.ts
@@ -57,6 +57,8 @@ import { apiSlice as partnerNetworkApiSlice } from './newPartnerNetwork/partnerN
 import { apiSlice as partnerNetworkPortalApiSlice } from './newPartnerNetwork/partnerNetworkPortalApiSlice'
 import { apiSlice as connectorApiSlice } from './connector/connectorApiSlice'
 import { apiSlice as serviceManagementApiSlice } from './serviceManagement/apiSlice'
+import serviceSubscriptionSlice from './serviceSubscription/slice'
+import { apiSlice as serviceSubscriptionApiSlice } from './serviceSubscription/serviceSubscriptionApiSlice'
 
 export const reducers = {
   admin,
@@ -68,6 +70,7 @@ export const reducers = {
   serviceMarketplace: serviceMarketplaceSlice.reducer,
   serviceProvider: serviceProviderSlice.reducer,
   appSubscription: appSubscriptionSlice.reducer,
+  serviceSubscription: serviceSubscriptionSlice.reducer,
   adminBoard: adminBoardSlice.reducer,
   userAdd: userAddSlice.reducer,
   userRole: userRoleSlice.reducer,
@@ -97,6 +100,8 @@ export const reducers = {
     partnerNetworkPortalApiSlice.reducer,
   [connectorApiSlice.reducerPath]: connectorApiSlice.reducer,
   [serviceManagementApiSlice.reducerPath]: serviceManagementApiSlice.reducer,
+  [serviceSubscriptionApiSlice.reducerPath]:
+    serviceSubscriptionApiSlice.reducer,
 }
 
 export const store = configureStore({
@@ -120,7 +125,8 @@ export const store = configureStore({
       .concat(partnerNetworkApiSlice.middleware)
       .concat(partnerNetworkPortalApiSlice.middleware)
       .concat(connectorApiSlice.middleware)
-      .concat(serviceManagementApiSlice.middleware),
+      .concat(serviceManagementApiSlice.middleware)
+      .concat(serviceSubscriptionApiSlice.middleware),
 })
 
 type RootState = ReturnType<typeof store.getState>